+++ src/psygnal/__init__.py
@@ -50,7 +50,7 @@ if os.getenv("PSYGNAL_UNCOMPILED"):
 
 from ._evented_decorator import evented
 from ._exceptions import EmitLoopError
-from ._group2 import EmissionInfo, SignalGroup
+from ._group import EmissionInfo, SignalGroup
 from ._group_descriptor import (
     SignalGroupDescriptor,
     get_evented_namespace,
