+++ src/psygnal/_group.py
@@ -9,12 +9,14 @@ the args that were emitted.
 """
 from __future__ import annotations
 
+import warnings
 from typing import (
     Any,
     Callable,
     ClassVar,
     ContextManager,
     Iterable,
+    Iterator,
     Mapping,
     NamedTuple,
 )
@@ -22,8 +24,6 @@ from typing import (
 from mypy_extensions import mypyc_attr
 
 from psygnal._signal import Signal, SignalInstance, _SignalBlocker
-
-__all__ = ["EmissionInfo", "SignalGroup"]
 
 
 class EmissionInfo(NamedTuple):
@@ -39,98 +39,23 @@ class EmissionInfo(NamedTuple):
     args: tuple[Any, ...]
 
 
-@mypyc_attr(allow_interpreted_subclasses=True)
-class SignalGroup(SignalInstance):
-    """`SignalGroup` that enables connecting to all `SignalInstances`.
+class SignalRelay(SignalInstance):
+    """Special SignalInstance that can be used to connect to all signals in a group."""
 
-    Parameters
-    ----------
-    instance : Any, optional
-        An instance to which this event group is bound, by default None
-    name : str, optional
-        Optional name for this event group, by default will be the name of the group
-        subclass.  (e.g., 'Events' in the example below.)
-
-    Examples
-    --------
-    >>> class Events(SignalGroup):
-    ...     sig1 = Signal(str)
-    ...     sig2 = Signal(str)
-    ...
-    >>> events = Events()
-    ...
-    >>> def some_callback(record):
-    ...     record.signal  # the SignalInstance that emitted
-    ...     record.args    # the args that were emitted
-    ...
-    >>> events.connect(some_callback)
-
-    note that the `SignalGroup` may also be created with `strict=True`, which will
-    enforce that *all* signals have the same emission signature
-
-    This is ok:
-
-    >>> class Events(SignalGroup, strict=True):
-    ...     sig1 = Signal(int)
-    ...     sig1 = Signal(int)
-
-    This will raise an exception
-
-    >>> class Events(SignalGroup, strict=True):
-    ...     sig1 = Signal(int)
-    ...     sig1 = Signal(str)  # not the same signature
-    """
-
-    _signals_: ClassVar[Mapping[str, Signal]]
-    _uniform: ClassVar[bool] = False
-
-    # this is only here to indicate to mypy that all attributes are SignalInstances.
-    def __getattr__(self, name: str) -> SignalInstance:
-        raise AttributeError(
-            f"SignalGroup {type(self).__name__!r} has no attribute {name!r}"
-        )
-
-    def __init_subclass__(cls, strict: bool = False) -> None:
-        """Finds all Signal instances on the class and add them to `cls._signals_`."""
-        cls._signals_ = {}
-        for k in dir(cls):
-            v = getattr(cls, k)
-            if isinstance(v, Signal):
-                cls._signals_[k] = v
-
-        cls._uniform = _is_uniform(cls._signals_.values())
-        if strict and not cls._uniform:
-            raise TypeError(
-                "All Signals in a strict SignalGroup must have the same signature"
-            )
-
-        return super().__init_subclass__()
-
-    def __init__(self, instance: Any = None, name: str | None = None) -> None:
-        super().__init__(
-            signature=(EmissionInfo,),
-            instance=instance,
-            name=name or self.__class__.__name__,
-        )
+    def __init__(self, group: SignalGroup, instance: Any = None) -> None:
+        self._group = group
+        super().__init__(signature=(EmissionInfo,), instance=instance)
         self._sig_was_blocked: dict[str, bool] = {}
-        for _, sig in self.signals.items():
-            sig.connect(self._slot_relay, check_nargs=False, check_types=False)
 
-    def __len__(self) -> int:
-        return len(self._slots)
-
-    @property
-    def signals(self) -> dict[str, SignalInstance]:
-        """Return {name -> SignalInstance} map of all signal instances in this group."""
-        return {n: getattr(self, n) for n in type(self)._signals_}
-
-    @classmethod
-    def is_uniform(cls) -> bool:
-        """Return true if all signals in the group have the same signature."""
-        return cls._uniform
+        # silence any warnings about failed weakrefs (will occur in compiled version)
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            for sig in group._psygnal_instances.values():
+                sig.connect(self._slot_relay, check_nargs=False, check_types=False)
 
     def _slot_relay(self, *args: Any) -> None:
-        if emitter := Signal.current_emitter():
+        emitter = Signal.current_emitter()
+        if emitter:
             info = EmissionInfo(emitter, args)
             self._run_emit_loop((info,))
 
@@ -178,8 +103,8 @@ class SignalGroup(SignalInstance):
         """
 
         def _inner(slot: Callable) -> Callable:
-            for sig in self.signals.values():
-                sig.connect(
+            for sig in self._group:
+                self._group[sig].connect(
                     slot,
                     check_nargs=check_nargs,
                     check_types=check_types,
@@ -193,7 +118,8 @@ class SignalGroup(SignalInstance):
     def block(self, exclude: Iterable[str | SignalInstance] = ()) -> None:
         """Block this signal and all emitters from emitting."""
         super().block()
-        for k, v in self.signals.items():
+        for k in self._group:
+            v = self._group[k]
             if exclude and v in exclude or k in exclude:
                 continue
             self._sig_was_blocked[k] = v._is_blocked
@@ -202,9 +128,9 @@ class SignalGroup(SignalInstance):
     def unblock(self) -> None:
         """Unblock this signal and all emitters, allowing them to emit."""
         super().unblock()
-        for k, v in self.signals.items():
+        for k in self._group:
             if not self._sig_was_blocked.pop(k, False):
-                v.unblock()
+                self._group[k].unblock()
 
     def blocked(
         self, exclude: Iterable[str | SignalInstance] = ()
@@ -236,17 +162,101 @@ class SignalGroup(SignalInstance):
         ValueError
             If `slot` is not connected and `missing_ok` is False.
         """
-        for signal in self.signals.values():
-            signal.disconnect(slot, missing_ok)
+        for signal in self._group:
+            self._group[signal].disconnect(slot, missing_ok)
         super().disconnect(slot, missing_ok)
+
+
+@mypyc_attr(allow_interpreted_subclasses=True)
+# class SignalGroup(Mapping[str, SignalInstance]):
+class SignalGroup:
+    _signals_: ClassVar[Mapping[str, Signal]]
+    _uniform: ClassVar[bool] = False
+
+    # see comment in __init__.  This type annotation can be overriden by subclass
+    # to change the public name of the SignalRelay attribute
+    all: SignalRelay
+
+    def __init__(self, instance: Any = None) -> None:
+        cls = type(self)
+        if not hasattr(cls, "_signals_"):  # pragma: no cover
+            raise TypeError(
+                "Cannot instantiate SignalGroup directly.  Use a subclass instead."
+            )
+        self._psygnal_instances: dict[str, SignalInstance] = {
+            name: signal.__get__(self, cls) for name, signal in cls._signals_.items()
+        }
+        self._psygnal_relay = SignalRelay(self, instance)
+
+        # determine the public name of the signal relay.
+        # by default, this is "all", but it can be overridden by the user by creating
+        # a new name for the SignalRelay annotation on a subclass of SignalGroup
+        # e.g. `my_name: SignalRelay`
+        self._psygnal_relay_name = "all"
+        for base in cls.__mro__:
+            for key, val in getattr(base, "__annotations__", {}).items():
+                if val is SignalRelay:
+                    self._psygnal_relay_name = key
+                    break
+        setattr(self, self._psygnal_relay_name, self._psygnal_relay)
+
+    def __init_subclass__(cls, strict: bool = False) -> None:
+        """Finds all Signal instances on the class and add them to `cls._signals_`."""
+        cls._signals_ = {
+            k: val
+            for k, val in getattr(cls, "__dict__", {}).items()
+            if isinstance(val, Signal)
+        }
+
+        cls._uniform = _is_uniform(cls._signals_.values())
+        if strict and not cls._uniform:
+            raise TypeError(
+                "All Signals in a strict SignalGroup must have the same signature"
+            )
+        super().__init_subclass__()
+
+    # TODO: change type hint after completing deprecation of direct access to
+    # names on SignalRelay object
+    def __getattr__(self, name: str) -> Any:
+        if name != "_psygnal_instances":
+            if name in self._psygnal_instances:
+                return self._psygnal_instances[name]
+            if name != "_psygnal_relay" and hasattr(self._psygnal_relay, name):
+                #   TODO: add deprecation warning and redirect to `self.all`
+                return getattr(self._psygnal_relay, name)
+        raise AttributeError(f"{type(self).__name__!r} has no attribute {name!r}")
+
+    @property
+    def signals(self) -> Mapping[str, SignalInstance]:
+        # TODO: deprecate this property
+        return self._psygnal_instances
+
+    def __len__(self) -> int:
+        return len(self._signals_)
+
+    def __getitem__(self, item: str) -> SignalInstance:
+        return self._psygnal_instances[item]
+
+    def __iter__(self) -> Iterator[str]:
+        return iter(self._signals_)
 
     def __repr__(self) -> str:
         """Return repr(self)."""
-        name = f" {self._name!r}" if self._name else ""
-        instance = f" on {self.instance!r}" if self.instance else ""
-        nsignals = len(self.signals)
-        signals = f"{nsignals} signals" if nsignals > 1 else ""
-        return f"<SignalGroup{name} with {signals}{instance}>"
+        name = self.__class__.__name__
+        instance = ""
+        nsignals = len(self)
+        signals = f"{nsignals} signals"
+        return f"<SignalGroup {name!r} with {signals}{instance}>"
+
+    @classmethod
+    def is_uniform(cls) -> bool:
+        """Return true if all signals in the group have the same signature."""
+        # TODO: Deprecate this method
+        return cls._uniform
+
+    def __deepcopy__(self, memo: dict[int, Any]) -> SignalGroup:
+        # TODO: should we also copy connections?
+        return type(self)(instance=self._psygnal_relay.instance)
 
 
 def _is_uniform(signals: Iterable[Signal]) -> bool:
